Use index.js only to forward main component
I find it easier to navigate the code when the filenames match the component names
